requirements.txt
httpx==0.27.0
langchain==0.1.17
langchain_community==0.0.36
langchain_core==0.1.50
langchain_openai==0.1.6
numpy==1.24.4
openai==1.25.2
pandas==2.2.2
streamlit==1.34.0
streamlit_modal==0.1.2
sentence-transformers==2.2.2
faiss_cpu==1.7.4
pydantic==1.10.13
python-dotenv==0.19.1
